fix(#1491): the committed listing is NUL-split and read strictly, never lossily
The DD1 family in a name-matching position: committed_listing read `git ls-files`
stdout through String::from_utf8_lossy, which substitutes one U+FFFD per invalid
byte — so two DISTINCT tracked fixtures can converge onto ONE string, and the
committed set keys on those strings (a BTreeSet of (dir, file) per table), so the
second fixture would silently vanish from the census rather than be reported as a
conflict. Without -z the spelling also depended on core.quotePath.

Now `git ls-files -z` (unquoted, NUL-separated) plus a pure parse_listing that
refuses a non-UTF-8 path fail-closed. Its test demonstrates the convergence it
prevents: from_utf8_lossy(&[0xff]) == from_utf8_lossy(&[0xfe]).

The inline test module moved to golden_fixture_root_tests.rs (pure move) — the
new case pushed the file past the 1500-line test-file threshold (epic #1135).
